+++ crates/scheduler/src/call/clock.rs
@@ -17,7 +17,7 @@ use wasefire_applet_api::clock::Api;
 use wasefire_applet_api::clock::{self as api};
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-clock")]
-use wasefire_board_api::AppletMemory as _;
+use wasefire_board_api::applet::Memory as _;
 #[cfg(feature = "board-api-clock")]
 use wasefire_board_api::{self as board, clock::Api as _};