+++ crates/scheduler/src/call/crypto/ecdh.rs
@@ -19,9 +19,11 @@ use wasefire_applet_api::crypto::ecdh::Kind;
 use wasefire_applet_api::crypto::ecdh::{self as api, Api, Curve};
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "internal-board-api-crypto-ecdh")]
+use wasefire_board_api::applet::{Memory as _, MemoryExt as _};
+#[cfg(feature = "internal-board-api-crypto-ecdh")]
 use wasefire_board_api::crypto::ecdh::Api as _;
 #[cfg(feature = "internal-board-api-crypto-ecdh")]
-use wasefire_board_api::{self as board, AppletMemory as _, AppletMemoryExt as _, Support};
+use wasefire_board_api::{self as board, Support};
 use wasefire_error::{Code, Error};
 
 #[cfg(feature = "internal-board-api-crypto-ecdh")]