+++ crates/scheduler/src/call/crypto/gcm.rs
@@ -15,11 +15,11 @@
 use wasefire_applet_api::crypto::gcm::{self as api, Api};
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-crypto-aes256-gcm")]
+use wasefire_board_api::applet::{Memory as _, MemoryExt as _};
+#[cfg(feature = "board-api-crypto-aes256-gcm")]
 use wasefire_board_api::crypto::aead::Api as _;
 #[cfg(feature = "board-api-crypto-aes256-gcm")]
 use wasefire_board_api::{self as board, Support as _};
-#[cfg(feature = "board-api-crypto-aes256-gcm")]
-use wasefire_board_api::{AppletMemory as _, AppletMemoryExt as _};
 
 #[cfg(feature = "board-api-crypto-aes256-gcm")]
 use crate::Trap;