+++ crates/scheduler/src/call/store.rs
@@ -20,7 +20,7 @@ use wasefire_applet_api::store::Api;
 use wasefire_applet_api::store::{self as api};
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-storage")]
-use wasefire_board_api::{AppletMemory as _, AppletMemoryExt as _};
+use wasefire_board_api::applet::{Memory as _, MemoryExt as _};
 
 use crate::DispatchSchedulerCall;
 #[cfg(feature = "board-api-storage")]