+++ crates/scheduler/src/call/usb/ctap.rs
@@ -19,7 +19,7 @@ use wasefire_applet_api::usb::ctap::Api;
 use wasefire_board_api as board;
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-usb-ctap")]
-use wasefire_board_api::AppletMemoryExt as _;
+use wasefire_board_api::applet::MemoryExt as _;
 #[cfg(feature = "board-api-usb-ctap")]
 use wasefire_board_api::usb::ctap::{Api as _, Event};